requirements.txt
flask
flask-cors
flask-login
flask-migrate
flask_seeder
gunicorn
requests
beautifulsoup4
psycopg2
psycopg2-binary
bcrypt
python-dotenv